mlx-app-converter/manual-tests/streamlit-llm-fixture/pyproject.toml
[project]
name = "streamlit-llm-fixture"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "transformers",
    "torch",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
